+++ tests/database/test_database.py
@@ -9,12 +9,13 @@ import tempfile
 from collections.abc import Generator
 from pathlib import Path
 from shutil import copyfile
-from typing import Any
+from typing import Any, cast
 
 import pytest
 
 from cobib.config import LabelSuffix, config
 from cobib.database import Database, Entry
+from cobib.database.database import CacheError
 
 from .. import get_resource
 
@@ -52,9 +53,9 @@ def setup() -> Generator[Any, None, None]:
         Access to the local fixture variables.
     """
     config.load(get_resource("debug.py"))
-    yield
-    Database().clear()
     Database().read()
+    yield
+    Database.reset()
     config.defaults()
 
 
@@ -63,6 +64,17 @@ def test_database_singleton() -> None:
     bib = Database()
     bib2 = Database()
     assert bib is bib2
+
+
+def test_database_reset() -> None:
+    # pylint: disable=protected-access
+    """Test the Database.reset method."""
+    bib = Database()
+    assert len(bib.keys()) > 0
+    assert bib._read
+    Database.reset()
+    assert len(bib.keys()) == 0
+    assert not bib._read
 
 
 def test_database_missing_file(caplog: pytest.LogCaptureFixture) -> None:
@@ -250,3 +262,110 @@ def test_database_save_delete() -> None:
     finally:
         os.remove(config.database.file)
         config.database.file = EXAMPLE_LITERATURE
+
+
+def test_database_caching_disabled(caplog: pytest.LogCaptureFixture) -> None:
+    """Tests that the caching mechanism can be disabled.
+
+    Args:
+        caplog: the built-in pytest fixture.
+    """
+    config.database.cache = None
+
+    Database.reset()
+    Database.read()
+
+    assert (
+        "cobib.database.database",
+        10,
+        "Encountered the following exception during cache lookup: "
+        "'caching is disabled via the configuration'",
+    ) in caplog.record_tuples
+
+
+def test_database_cache_not_found() -> None:
+    """Test the Database.read_cache method."""
+    with tempfile.TemporaryDirectory() as tempdir:
+        config.database.cache = tempdir
+
+        Database.reset()
+
+        with pytest.raises(CacheError, match="the database has not been cached yet"):
+            Database.read_cache()
+
+
+def test_database_save_cache() -> None:
+    """Test the Database.save_cache method."""
+    with tempfile.TemporaryDirectory() as tempdir:
+        config.database.cache = tempdir
+
+        Database.save_cache()
+
+        with pytest.raises(OSError, match="Directory not empty"):
+            Path(tempdir).rmdir()
+
+
+def test_database_cache_uptodate(caplog: pytest.LogCaptureFixture) -> None:
+    """Test Database.save_cache stops when the cache is up-to-date.
+
+    Args:
+        caplog: the built-in pytest fixture.
+    """
+    with tempfile.TemporaryDirectory() as tempdir:
+        config.database.cache = tempdir
+
+        Database.save_cache()
+
+        Database.save_cache()
+
+    assert (
+        "cobib.database.database",
+        20,
+        "The cache is already up-to-date",
+    ) in caplog.record_tuples
+
+
+def test_database_cache_reading(caplog: pytest.LogCaptureFixture) -> None:
+    """Test that the Database.read method reads the cache.
+
+    Args:
+        caplog: the built-in pytest fixture.
+    """
+    with tempfile.TemporaryDirectory() as tempdir:
+        config.database.cache = tempdir
+
+        Database.save_cache()
+        Database.reset()
+        Database.read()
+
+    for record in caplog.record_tuples:
+        if record[0] != "cobib.database.database":
+            continue
+        if record[1] != 10:
+            continue
+        if "Reading the cached database from " in record[2]:
+            break
+    else:
+        pytest.fail("Log message not found which would indicate that Database.read_cache worked")
+
+
+def test_database_cache_outdated() -> None:
+    """Test Database.read_cache stops when the cache is outdated."""
+    with tempfile.TemporaryDirectory() as tempdir:
+        config.database.cache = tempdir
+        config.database.file = Path(tempdir) / "cobib_test_database_file.yaml"
+        copyfile(EXAMPLE_LITERATURE, config.database.file)
+
+        Database.reset()
+        Database.read()  # this will trigger save_cache, too
+
+        # pylint: disable=protected-access
+        cache_file = Database._get_cache_file()
+        new_time = cast(Path, cache_file).stat().st_mtime_ns
+        # make the cache appear outdated by updating the modified time of the database file
+        os.utime(config.database.file, ns=(new_time + 1_000_000, new_time + 1_000_000))
+
+        Database.reset()
+
+        with pytest.raises(CacheError, match="the cached database is outdated"):
+            Database.read_cache()