+++ tests/debug.py
@@ -7,6 +7,7 @@ from cobib.config import config
 config.logging.version = None
 
 root = Path(__file__).parent
+config.database.cache = None
 config.database.file = str((root / "example_literature.yaml").resolve())
 
 config.utils.file_downloader.url_map[